requirements.txt
# Core Dependencies
yt-dlp>=2024.1.0
customtkinter>=5.2.0
pillow>=10.0.0
requests>=2.31.0

# Media Processing
ffmpeg-python>=0.2.0

# Additional Features
selenium>=4.15.0
webdriver-manager>=4.0.0

# Optional but recommended
pyinstaller>=6.0.0  # For building executable
